+++ app/services/expired_registrations_service.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class ExpiredRegistrationsService < ::WasteExemptionsEngine::BaseService
+class ExpiredRegistrationsService < WasteExemptionsEngine::BaseService
   def run
     all_expired_registration_exemptions.each(&:expire!)
   end