+++ config/initializers/defra_ruby_aws.rb
@@ -24,6 +24,16 @@ DefraRuby::Aws.configure do |c|
     encrypt_with_kms: ENV.fetch("AWS_DAILY_ENCRYPT_WITH_KMS", nil)
   }
 
+  deregistration_email_export_bucket = {
+    name: ENV.fetch("AWS_DEREGISTRATION_EMAIL_EXPORT_BUCKET", nil),
+    region: ENV.fetch("AWS_REGION", nil),
+    credentials: {
+      access_key_id: ENV.fetch("AWS_DEREGISTRATION_EMAIL_EXPORT_ACCESS_KEY_ID", nil),
+      secret_access_key: ENV.fetch("AWS_DEREGISTRATION_EMAIL_EXPORT_SECRET_ACCESS_KEY", nil)
+    },
+    encrypt_with_kms: ENV.fetch("AWS_DEREGISTRATION_EMAIL_ENCRYPT_WITH_KMS", nil)
+  }
+
   boxi_export_bucket = {
     name: ENV.fetch("AWS_BOXI_EXPORT_BUCKET", nil),
     region: ENV.fetch("AWS_REGION", nil),
@@ -34,6 +44,11 @@ DefraRuby::Aws.configure do |c|
     encrypt_with_kms: ENV.fetch("AWS_BOXI_ENCRYPT_WITH_KMS", nil)
   }
 
-  c.buckets = [bulk_bucket, epr_bucket, boxi_export_bucket]
+  c.buckets = [
+    bulk_bucket,
+    epr_bucket,
+    boxi_export_bucket,
+    deregistration_email_export_bucket
+  ]
 end
 # rubocop:enable Metrics/BlockLength