+++ config/initializers/defra_ruby_features.rb
@@ -4,5 +4,5 @@ require "defra_ruby_features"
 
 DefraRubyFeatures.configure do |configuration|
   # Tell the engine where to find a model to use in order to persist feature toggles data
-  configuration.feature_toggle_model_name = "::WasteExemptionsEngine::FeatureToggle"
+  configuration.feature_toggle_model_name = "WasteExemptionsEngine::FeatureToggle"
 end