+++ spec/factories/renewing_registration.rb
@@ -3,10 +3,14 @@
 FactoryBot.define do
   factory :renewing_registration, class: "WasteExemptionsEngine::RenewingRegistration" do
     # Create a new registration when initializing so we can copy its data
-    initialize_with do
-      registration = create(:registration)
+    registration { create(:registration) }
 
+    initialize_with do
       new(reference: registration.reference, token: registration.renew_token)
+    end
+
+    trait :expires_tomorrow do
+      registration { create(:registration, :expires_tomorrow) }
     end
   end
 end